requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyPDF2==3.0.1
pytesseract==0.3.10
numpy==1.24.3
scipy==1.10.1
sympy==1.12
pandas==2.0.3
matplotlib==3.7.1
python-multipart==0.0.6
Pillow==10.0.0
latex2sympy2==1.9.1
antlr4-python3-runtime==4.11.1
mistralai>=1.0.0
aiohttp==3.9.1 